+++ src/KurrentDB.SecondaryIndexing/Indexes/Category/CategoryIndexReader.cs
@@ -26,10 +26,10 @@ class CategoryIndexReader(
 		return processor.GetCategoryId(categoryName);
 	}
 
-	protected override long GetLastIndexedSequence(long id) => processor.GetLastEventNumber(id);
+	protected override long GetLastIndexedSequence(long id) => processor.GetLastEventNumber((int)id);
 
 	protected override IEnumerable<IndexedPrepare> GetIndexRecords(long id, long fromEventNumber, long toEventNumber) {
-		var range = db.Pool.Query<CategoryIndexQueryArgs, CategoryRecord, CategoryIndexQuery>(new(id, fromEventNumber, toEventNumber));
+		var range = db.Pool.Query<CategoryIndexQueryArgs, CategoryRecord, CategoryIndexQuery>(new((int)id, fromEventNumber, toEventNumber));
 		if (range.Count < toEventNumber - fromEventNumber + 1) {
 			// events might be in flight
 			var inFlight = queryInFlightRecords(
